requirements.txt
slack_bolt
yfinance
pandas
numpy
python-dotenv
